Extract requestChainxInfo
