+++ src/pages/NodeAction/index.js
@@ -18,6 +18,7 @@ import {
 import PrimaryButton from '@chainx/ui/dist/Button/PrimaryButton'
 import InfoLabel from '../../components/InfoLabel'
 import { sleep } from '../../shared'
+import requestChainxInfo from '@pages/NodeAction/requestChainxInfo'
 
 function AddNode(props) {
   const [name, setName] = useState('')
@@ -27,45 +28,6 @@ function AddNode(props) {
   const dispatch = useDispatch()
   const isTestNet = useSelector(isTestNetSelector)
 
-  const requestChainInfo = url => {
-    const id = Number(
-      Date.now() +
-        Math.random()
-          .toString()
-          .substr(2, 3)
-    ).toString(36)
-
-    const requestMsg = JSON.stringify({
-      id,
-      jsonrpc: '2.0',
-      method: 'system_properties',
-      params: []
-    })
-
-    return new Promise((resolve, reject) => {
-      const ws = new WebSocket(url)
-
-      ws.onopen = () => {
-        ws.send(requestMsg)
-      }
-
-      ws.onmessage = res => {
-        const data = JSON.parse(res.data)
-        if (data.id !== id) {
-          return
-        }
-
-        resolve({ data: data.result })
-        ws.close()
-      }
-
-      ws.onerror = err => {
-        reject({ err })
-        ws.close()
-      }
-    })
-  }
-
   const enter = async () => {
     if (!name || !url) {
       setErrMsg('name and url are required')
@@ -73,13 +35,15 @@ function AddNode(props) {
     }
 
     try {
-      const result = await Promise.race([requestChainInfo(url), sleep(10000)])
+      const result = await Promise.race([requestChainxInfo(url), sleep(10000)])
       if (!result) {
         setErrMsg('Fetch node properties timeout')
         return
       }
 
-      const { network_type: network } = result || {}
+      const {
+        data: { network_type: network }
+      } = result || {}
       if (
         (!isTestNet && network !== 'mainnet') ||
         (isTestNet && network !== 'testnet')
